### 2016-07-20 #40 @whedon
Submission: PsychoPhysioCollector
**Submitting author:** @sbogutzky (<a href="http://orcid.org/0000-0002-9279-6549">Simon Bogutzky</a>)
**Repository:** <a href="https://github.com/sbogutzky/PsychoPhysioCollector" target ="_blank">https://github.com/sbogutzky/PsychoPhysioCollector</a>
**Version:** v2.0.5
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.59387" target="_blank" >10.5281/zenodo.59387</a>
