### 2016-07-29 #40 @arfon
@openjournals/joss-editors - any suggestions for potential reviewers here?
