### 2016-07-29 #40 @davclark
So you just need someone to answer the reviewer questions on this page for the PDF? Here on this issue?
