### 2016-07-29 #40 @arfon
@davclark - basically yes. There's an associated GitHub repository (https://github.com/sbogutzky/PsychoPhysioCollector) which is the main focus of this review together with the short PDF. See our reviewer guidelines here http://joss.theoj.org/about#reviewer_guidelines

If you think you might be able to help out, I'll add you to the `@openjournals/joss-reviewers` team so that you have permissions to update the checkboxes at the top of this issue.
