### 2016-07-29 #40 @arfon
@davclark 🎉 . You should have an invite to join the '@joss-reviewers' [team]
(https://github.com/orgs/openjournals/invitation?via_email=1) in your inbox. [This link](https://github.com/orgs/openjournals/invitation?via_email=1) should also work.

Please shout if anything isn't clear about the review process here.
