### 2016-08-01 #40 @davclark
@arfon do we verify that the license is enforceable? I.e., that the copyright holder indeed has copyright?
