### 2016-08-01 #40 @davclark
Cool stuff. This is important for social science research - and warrants a change in the language above from the JOSS review checklist:

> (ideally to solve real-world analysis problems)

*Data collection* is also part of science - much harder than analysis in many experiments. To this end, other guidelines are strained. What is the significance of "API documentation?"

A few more notes:

- The tag 2.0.4 does not include the paper.
- I cannot verify all functionality of this software as it requires specialized hardware. I do have an android phone.
- I don't think I have time to install android development tooling. Is there a ready-made APK that I can *directly* access via the web for installation on my phone? I know I can use android file transfer, but this is a barrier for some users (and would make my life faster too).

- I would clarify that anyone can install the existing APK easily, but that an android development environment is needed to create new questionnaires. In the paper proper, it would also be good to clarify that a few self-report surveys are provided, and that more can be created with proper android development tools.
- Link to android developer tools / docs?
- Could a brief developer doc be provided that points to top-level classes / files for major functionality? This would greatly help with contributions.
- There is no advise on getting help or contributing, except that you are welcome.
- Please spell out acronyms. These include ESM, IMU and ECG.
- Your audience is almost certainly larger than people running "ESM protocols." Please provide a general level of technical aptitude required to use the app *and* get the data off an analyze it. Give a general sense of quality of biosignals.
- What are existing solutions like? Why wouldn't I just use them?
- Could you say a little more about how "its visualization component allows checking the experimental setup in real time." This would address the question of manual steps to verify operation.
- Is there more written about Flow-Machines?
- The paper appears to be missing the required link to the repo.
