### 2016-08-01 #40 @davclark
While the above seems like a lot, I'd call this "minor revisions" modulo some clarification on the things that break the "analysis software" model - most critically that I can't actually test this on the biosensor hardware.
