### 2016-08-01 #40 @arfon
> @arfon do we verify that the license is enforceable? I.e., that the copyright holder indeed has copyright?

@davclark - If you know of a reason why the author may have go this wrong then I think it's worth raising the question but we don't want to give legal advice to those submitting.

> @whedon I see you've copyrighted the work to yourself. If produced at an american university, this is often not legal - we need to assign copyright to the institution if we worked on it as a student. For free software, this is usually not a big deal - especially for a very liberal license.

@davclark - @whedon is actually the bot that made the submission issue. @sbogutzky - this question is for you 😉
