### 2016-08-01 #40 @arfon
> (ideally to solve real-world analysis problems)

> Data collection is also part of science - much harder than analysis in many experiments. To this end, other guidelines are strained. What is the significance of "API documentation?"

@davclark agreed and thanks for raising this point. Would you be willing to open an issue on https://github.com/openjournals/joss outlining your concerns with the language used in the checklist?
