### 2016-08-01 #40 @sbogutzky
> The tag 2.0.4 does not include the paper.

Yes, this is right. tag 2.0.4 is the latest version of the PPC without the paper.

> I cannot verify all functionality of this software as it requires specialized hardware. I do have an android phone.

But you can test the basic functionality, because data will always collect from the internal sensors of the smartphone. In this way, you can for example track the arm motion of a participant with a sport armband while running.

> I don't think I have time to install android development tooling. Is there a ready-made APK that I can directly access via the web for installation on my phone? I know I can use android file transfer, but this is a barrier for some users (and would make my life faster too).

> I would clarify that anyone can install the existing APK easily, but that an android development environment is needed to create new questionnaires.

So, you find the APK. See Installation Instructions of the readme.md

> In the paper proper, it would also be good to clarify that a few self-report surveys are provided, and that more can be created with proper android development tools.

Ok.

> Link to android developer tools / docs?

See Used Libraries in the readme.md

> Could a brief developer doc be provided that points to top-level classes / files for major functionality? This would greatly help with contributions.
> There is no advise on getting help or contributing, except that you are welcome.

> Please spell out acronyms. These include ESM, IMU and ECG.

Ok.

> Your audience is almost certainly larger than people running "ESM protocols." Please provide a general level of technical aptitude required to use the app and get the data off an analyze it. Give a general sense of quality of biosignals.

Ok, but this depends on the sensors and the configuration, which you use.

> What are existing solutions like? Why wouldn't I just use them?

Maybe there are other solution, but I these solutions do not support the hardware of our equipment.

> Could you say a little more about how "its visualization component allows checking the experimental setup in real time." This would address the question of manual steps to verify operation.

E. g. if you place electrodes on the human body, you can check the leakage of the ECG based on the realtime visualization of the heart activity and the pattern of the signal.

> Is there more written about Flow-Machines?
Yes, of course, unfortunately the most is in german. There one paper of the beginning: "Gait Biometrics in Interaction Design: A Work in Progress Report" and presentation: "Analysis of ‚flow while walking and running’: psychophysiological correlates and activity-based results".
In this year we will also published some scientific papers about our research in the project.

> The paper appears to be missing the required link to the repo.
Is that needed?
