### 2016-08-01 #40 @davclark
I think I may have inter-mixed some different kinds of issues above, so I'll respond to a few things again here. Let me also add that assuming it works (as I do), this is a very useful contribution. My comments are simply directed at making the publication / documentation better, and opening this up to more potential users and perhaps developers.

## Things that are up to the editors

The things in this section will have no bearing on my recommendation.

>> The tag 2.0.4 does not include the paper.

> Yes, this is right. tag 2.0.4 is the latest version of the PPC without the paper.

I'm not sure if there are rules here, and I'll leave this to the editors. Seems better to have the paper in the tagged release, though.

>> The paper appears to be missing the required link to the repo.

> Is that needed?

This is in the guidelines for reviewers (emphasis mine):

> JOSS papers should contain the following:

> ...

> - A list of key references **including a link to the software archive**

As a reviewer, I feel my job is done by pointing out that this is not done.

## Things that are "mine"

>> I cannot verify all functionality of this software as it requires specialized hardware. I do have an android phone.

> But you can test the basic functionality, because data will always collect from the internal sensors of the smartphone. In this way, you can for example track the arm motion of a participant with a sport armband while running.

I've opened openjournals/joss#156 on this point. Given the current discussion, I think we're gonna let the things I can't test slide.

>> I don't think I have time to install android development tooling. Is there a ready-made APK that I can directly access via the web for installation on my phone? I know I can use android file transfer, but this is a barrier for some users (and would make my life faster too).

>> I would clarify that anyone can install the existing APK easily, but that an android development environment is needed to create new questionnaires.

>So, you find the APK. See Installation Instructions of the readme.md

It seems like there is either a disagreement or lack of understanding here. Let me give a more concrete example:

I am at work with my fancy/annoying new macbook retina. To connect my phone, I need to unplug the hub from the single USB-C port, which means I need to rearrange my desk and move my ergo keyboard out of the way so I can plug into the USB-C port on my phone. I then grab the cable from the power adapter to connect my phone to my computer. The future is going to look more like this for more people.

What I'm suggesting is easy enough that I just installed the APK by creating a
fork and a new release [here](https://github.com/davclark/PsychoPhysioCollector/releases/tag/2.0.4). Just like what you have, except including a direct download of the APK.

>> Link to android developer tools / docs?

> See Used Libraries in the readme.md

Those don't include https://developer.android.com/studio/index.html

Compiling this would not be hard for most people with rudimentary development skills. You are not currently giving the basic pieces that would facilitate this.

>> What are existing solutions like? Why wouldn't I just use them?

> Maybe there are other solutions, but I these solutions do not support the hardware of our equipment.

Here is [another solution](http://www.shimmersensing.com/shop/neurolynq/). Why wouldn't I just use that? I can at least partially answer this question, but I'm asking you to at a minimum cite primary existing solutions *in your paper* per JOSS guidelines.

>> Could you say a little more about how "its visualization component allows checking the experimental setup in real time." This would address the question of manual steps to verify operation.

> E. g. if you place electrodes on the human body, you can check the leakage of the ECG based on the realtime visualization of the heart activity and the pattern of the signal.

It seems there is no similar readout for phone sensors? Or am I missing something?

I am able to verify the contents of the CSV files manually. Note that there is a slight type-o in your documentation. You mention `psychophysiocollector/PARTICIPANT_NAME/ACTIVITY_NAME`, but you actually get them in `psychophysiocollector/ACTIVITY_NAME/PARTICIPANT_NAME/DATETIME`. Maybe just say `PsychoPhysioCollector` and leave the rest to the user (it's obvious from there).

>> Is there more written about Flow-Machines?

> Yes, of course, unfortunately the most is in german. There one paper of the beginning: "Gait Biometrics in Interaction Design: A Work in Progress Report" and presentation: "Analysis of ‚flow while walking and running’: psychophysiological correlates and activity-based results".
> In this year we will also published some scientific papers about our research in the project.

You should cite these - It doesn't hurt anything, and it facilitates discovery of your work! But I won't demand it.
