### 2016-08-02 #40 @sbogutzky
@davclark - First of all, that's for your feedback.

@arfon - Now some questions to the editors.

> The tag 2.0.4 does not include the paper.
>> Yes, this is right. tag 2.0.4 is the latest version of the PPC without the paper.
>>> I'm not sure if there are rules here, and I'll leave this to the editors. Seems better to have the paper in the tagged release, though.

Is there a rule?

> The paper appears to be missing the required link to the repo.
>> Is that needed?
>>> This is in the guidelines for reviewers (emphasis mine):

>>> JOSS papers should contain the following:

>>> ...

>>> A list of key references including a link to the software archive
>>> As a reviewer, I feel my job is done by pointing out that this is not done.

I do not get it, sorry? Should I at the repo as webpage to the bib file and reference the entry in the paper? I do not see this in other paper of the journal.

There, I see Paper DOI; Software repository and Software Archive. But these things will be added by the editors, right? After accepting?
