### 2016-08-02 #40 @sbogutzky
@davclark -
> It seems like there is either a disagreement or lack of understanding here. Let me give a more concrete example:
>
> I am at work with my fancy/annoying new macbook retina. To connect my phone, I need to unplug the hub from the single USB-C port, which means I need to rearrange my desk and move my ergo keyboard out of the way so I can plug into the USB-C port on my phone. I then grab the cable from the power adapter to connect my phone to my computer. The future is going to look more like this for more people.
>
> What I'm suggesting is easy enough that I just installed the APK by creating a
> fork and a new release here. Just like what you have, except including a direct download of the APK.

Sorry, there is a lack of understanding. What is the difference of my APK and your fork? In my case, I downloaded the APK from the repo and put it in my Google Drive or Dropbox folder and installed it on the smartphone.
