### 2016-08-02 #40 @sbogutzky
@davclark -

> What are existing solutions like? Why wouldn't I just use them?
>> Maybe there are other solutions, but I these solutions do not support the hardware of our equipment.
>> Here is another solution. Why wouldn't I just use that? I can at least partially answer this question, but I'm asking you to at a minimum cite primary existing solutions in your paper per JOSS guidelines.

Ok. Right. This is a quite new solution of Shimmer for a newer type of IMUs. We began our studies in 2013 and build the existing solution for an older type of Shimmer IMUs (R2). These solution seems not to meet our requirements of mobility and automatic time synchronization of more data streams. A solution, which can you use, if you find it, is that: "Psychophysiological Correlates of Flow During Daily Activities" by Gaggioli et al. (2013)

>>> Could you say a little more about how "its visualization component allows checking the experimental setup in real time." This would address the question of manual steps to verify operation.
>> E. g. if you place electrodes on the human body, you can check the leakage of the ECG based on the realtime visualization of the heart activity and the pattern of the signal.
> It seems there is no similar readout for phone sensors? Or am I missing something?

No, because we worked in our studies primary with the Shimmer R2 IMUs.

> I am able to verify the contents of the CSV files manually. Note that there is a slight type-o in your documentation. You mention psychophysiocollector/PARTICIPANT_NAME/ACTIVITY_NAME, but you actually get them in psychophysiocollector/ACTIVITY_NAME/PARTICIPANT_NAME/DATETIME. Maybe just say PsychoPhysioCollector and leave the rest to the user (it's obvious from there).

That is right, I fixed it!
