### 2016-08-02 #40 @davclark
@sbogutzky I am pretty sure at this point that I've communicated my concerns / recommendations - so please make any remaining revisions and I'll wait for you to tell me you're done. I'll also wait for @arfon or another editor to chime in on the questions for them.

Regarding the google drive solution, that makes good sense. You've illustrated that there is a pretty easy way for individuals to install the APK without a wire. My solution was the same APK as yours, just providing a direct link from the github release page. I *still* think that'd be a nice-to-have, but I don't think this is required.
