### 2016-08-03 #40 @sbogutzky
@davclark - Thank for your recommendations. I'm ready with my revision.

[![DOI](https://zenodo.org/badge/23671/sbogutzky/PsychoPhysioCollector.svg)](https://zenodo.org/badge/latestdoi/23671/sbogutzky/PsychoPhysioCollector)
