### 2016-08-03 #40 @davclark
Looks good to me, and I hope the changes are helpful to others and (perhaps more importantly) for you substantiating this work as a genuine academic contribution.

So as reviewer, I recommend acceptance. It's up to the editors about the verification issue described above (that I can't verify all functionality without the sensors), but I recommend the "low bar" of accepting based on verification "as possible" and trusting the authors for the rest (including evidence from the provided references in which this software was used with the sensors).

Editors, please also note that the tag is updated to 2.0.5.
