### 2016-08-03 #40 @sbogutzky
Editors, please also note that change the title to "PsychoPhysioCollector: A Smartphone-Based Data Collection App for Psychophysiological Research"

It is possible to change that on the Paper page?
