### 2016-08-04 #40 @arfon
Thanks for the review @davclark and for being responsive @sbogutzky. Some responses to questions that have come up in this thread:

> Is there a rule? (about the tagged releases)

Ideally the release should include the paper but I don't think this is a hard rule right now. More important to me is that the release matches the version that has been archived (e.g. with Zenodo).

> There, I see Paper DOI; Software repository and Software Archive. But these things will be added by the editors, right? After accepting?

That's correct. The final version of the PDF will include a direct link to the software repository and the archive of the code.

> Editors, please also note that change the title to "PsychoPhysioCollector: A Smartphone-Based Data Collection App for Psychophysiological Research"

This is now changed.

> So as reviewer, I recommend acceptance. It's up to the editors about the verification issue described above (that I can't verify all functionality without the sensors), but I recommend the "low bar" of accepting based on verification "as possible" and trusting the authors for the rest (including evidence from the provided references in which this software was used with the sensors).

👍
