### 2016-08-04 #40 @arfon
@sbogutzky you paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00040 🎉 🚀 💥

Thanks for all of your help @davclark!
